fix(build): resolve C# API incompatibilities for CUE4Parse/CefGlue downgrades
Adapt code to CUE4Parse 1.2.2 API (different namespaces, missing methods),
CefGlue 0.4.0 API (simplified handler signatures), and fix UniFFI bindings
to compile against available types.

+++ godot/scripts/MainController.cs
@@ -289,7 +289,7 @@ public partial class MainController : Node
                 EventType = Xilium.CefGlue.CefKeyEventType.Char,
                 Character = (char)key.Unicode,
                 UnmodifiedCharacter = (char)key.Unicode,
-                WindowsKeyCode = key.Unicode,
+                WindowsKeyCode = (int)key.Unicode,
                 Modifiers = GetCefModifiers(key),
             };
             _browser.SendKeyEvent(charEvent);
